requirements.txt
altgraph==0.17.5
bidict==0.23.1
blinker==1.9.0
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
eventlet==0.40.3
Flask==3.1.2
Flask-SocketIO==5.5.1
greenlet==3.2.4
h11==0.16.0
hanzidentifier==1.3.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
langdetect==1.0.9
lxml==6.0.2
MarkupSafe==3.0.3
minimap-renderer @ git+https://github.com/WoWs-Builder-Team/minimap_renderer.git@8ce5bab5ebe6dc3b6b394bfa6f90e5348db2ee31
numpy==2.3.5
packaging==25.0
pefile==2023.2.7
pillow==12.0.0
pycryptodomex==3.23.0
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
python-engineio==4.12.3
python-socketio==5.14.3
pywin32-ctypes==0.2.3
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
tqdm==4.67.1
waitress==3.0.2
Werkzeug==3.1.3
wheel==0.45.1
wsproto==1.3.2
zhon==2.1.1
